% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{are_zero}
\alias{are_zero}
\title{Checks if a vector is non-negative numeric}
\usage{
are_zero(x)
}
\arguments{
\item{x}{object to be checked}
}
\value{
vector logical of same length as x
}
\description{
Checks if a vector is non-negative numeric
}
\seealso{
Other predicates: \code{\link{are_non_negative}},
  \code{\link{are_numeric_like}},
  \code{\link{are_positive}},
  \code{\link{are_whole_number}},
  \code{\link{has_zero_range}},
  \code{\link{is_binary_valued}},
  \code{\link{is_scalar_bare_numeric}},
  \code{\link{is_zero}}
}
