% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.character.formula}
\alias{as.character.formula}
\alias{as.character}
\title{as.character.formula}
\usage{
\method{as.character}{formula}(x, ...)
}
\arguments{
\item{x}{formula object}

\item{...}{arguments passed to and from other methods.}
}
\value{
A character vector
}
\description{
Converts a formula to character representaion.
}
