% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzid.R
\name{as.data.frame.mzid}
\alias{as.data.frame.mzid}
\alias{as.data.frame}
\title{Prints an mzid object as a data frame}
\usage{
\method{as.data.frame}{mzid}(x, ...)
}
\arguments{
\item{x}{mzid object}

\item{...}{objects passed to \code{\link[dplyr]{as_data_frame}}}
}
\description{
Prints an mzid object as a data frame
}
