% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ckd_epi}
\alias{ckd_epi}
\title{ckd_epi}
\usage{
ckd_epi(gender, age, scr, black)
}
\arguments{
\item{gender}{numeric vector : 1 = male, 0 = female}

\item{age}{numeric vector : Age in years}

\item{scr}{numeric vector : Serum creatinine in mg/dl}

\item{black}{numeric vector : 1 = black, 0 = otherwise}
}
\value{
numeric vector : ckd_epi scores
}
\description{
Calculate eGFR using the CKD-EPI (Chronic Kidney Disease Epidemiology
Collaboration) formula.
\url{https://www.qxmd.com/calculate/calculator_251/egfr-using-ckd-epi}
}
