% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross.R
\name{cross_some}
\alias{cross_some}
\title{Crosses some vectors in a list, splicing other elements into each cross Useful for generating lists of substitutions}
\usage{
cross_some(.l, .at)
}
\arguments{
\item{.l}{list of vectors}

\item{.at}{A character vector of names or a numeric vector of positions}
}
\value{
list
}
\description{
Crosses some vectors in a list, splicing other elements into each cross Useful for generating lists of substitutions
}
