% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute.R
\name{cross_sub}
\alias{cross_sub}
\alias{f_cross_sub}
\title{Crosses a list of quoted calls with a list of environments, performing quoted
substitution each time.}
\usage{
cross_sub(exprs, envs, trans = identity)

f_cross_sub(fos, envs, trans = identity)
}
\arguments{
\item{exprs}{a list of quoted calls}

\item{envs}{a list of environments, or things that behaves like an
environment (like a list or data frame), or references to an environment
(like a positive integer or name, see \code{\link{as.environment}} for more
details)}

\item{trans}{a function to be applied to each value in an environment before
it is passed to `substitute_q`. A common vlaue is \code{as.name}, to
substitute in character values as names.}

\item{fos}{a list of formulas}
}
\value{
a nested list of calls
}
\description{
Useful for generating lists of model formulas. `f_cross_sub` is a version
that enforces formulas.
}
\details{
f_cross_sub maps environments across created calls.
}
