% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dat_to_mat}
\alias{dat_to_mat}
\title{Converts a dataframe to a matrix where a column of the dataframe becomes the
rownames of the matrix}
\usage{
dat_to_mat(x, var = 1)
}
\arguments{
\item{x}{data frame}

\item{var}{scalar tidyeval var: column to select. Should not contain a
negative (drop column).}
}
\value{
matrix
}
\description{
This is mostly a convenience function for conveting dataframes for clustering
like functions, or anything that needs a matrix.
}
\keyword{internal}
