% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deparse2}
\alias{deparse2}
\title{Expression Deparsing}
\usage{
deparse2(expr, ...)
}
\arguments{
\item{expr}{any R expression}

\item{...}{arguments passed to \code{\link{deparse}}}
}
\value{
scalar? character
}
\description{
Turn unevaluated expressions into character strings. With additional
operations to make it a little neater.
}
