% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strs.R
\name{deparse_to_str}
\alias{deparse_to_str}
\title{Deparse an object to a string}
\usage{
deparse_to_str(x)
}
\arguments{
\item{x}{object}
}
\value{
scalar? character
}
\description{
Mostly used for pulling language and call objects to strings.
}
