% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fo_to_char}
\alias{fo_to_char}
\alias{as.character.formula}
\alias{as.character}
\title{fo_to_char}
\usage{
fo_to_char(x, ...)

\method{as.character}{formula}(x, ...)
}
\arguments{
\item{x}{formula}

\item{...}{arguments passed to and from other methods.}
}
\value{
A character vector
}
\description{
Converts a formula to character representaion.
}
