% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Not In}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{vector or NULL. Values to be matched.}

\item{table}{vector or NULL. Values to be matched against.}
}
\value{
logical vector of same length as x, indicating if a match was located
  for each element of x
}
\description{
Not In
}
