% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_which}
\alias{inv_which}
\title{Title}
\usage{
inv_which(x, sel)
}
\arguments{
\item{x}{lits or atomic vector}

\item{sel}{A character vector of names or a numeric vector of positions}
}
\value{
vector logical
}
\description{
Title
}
\keyword{internal}
