% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_non_negative_numeric}
\alias{is_non_negative_numeric}
\title{Checks if a vector is non-negative numeric}
\usage{
is_non_negative_numeric(x, n = NULL)
}
\arguments{
\item{x}{object to be checked}

\item{n}{scalar wholenumber: expected length of vector}
}
\value{
scalar logical
}
\description{
Checks if a vector is non-negative numeric
}
\seealso{
Other predicates: \code{\link{is_binary_valued}},
  \code{\link{is_numeric_like}},
  \code{\link{is_positive_numeric}},
  \code{\link{is_scalar_numeric}},
  \code{\link{is_wholenumber}}
}
