% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_numeric_like}
\alias{is_numeric_like}
\title{Checks if a an object can be converted to a number without error}
\usage{
is_numeric_like(x, ...)
}
\arguments{
\item{x}{object to be checked}

\item{...}{argument passed to \code{\link{anyNA}}}
}
\value{
scalar logical
}
\description{
Checks if a an object can be converted to a number without error
}
\examples{
is_numeric_like(c("5", "1.23", "-3.14", "hello", "?"))
}
\seealso{
Other predicates: \code{\link{is_binary_valued}},
  \code{\link{is_non_negative_numeric}},
  \code{\link{is_positive_numeric}},
  \code{\link{is_scalar_numeric}},
  \code{\link{is_wholenumber}}
}
