% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_scalar_bare_numeric}
\alias{is_scalar_bare_numeric}
\title{Bare type predicates}
\usage{
is_scalar_bare_numeric(x)
}
\arguments{
\item{x}{Object to be tested.}
}
\description{
These predicates check for a given type but only return TRUE for bare R
objects. Bare objects have no class attributes. For example, a data frame is a
list, but not a bare list.
}
\seealso{
Other predicates: \code{\link{are_non_negative}},
  \code{\link{are_numeric_like}},
  \code{\link{are_positive}},
  \code{\link{are_whole_number}}, \code{\link{are_zero}},
  \code{\link{has_zero_range}},
  \code{\link{is_binary_valued}}, \code{\link{is_zero}}
}
