% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_scalar_numeric}
\alias{is_scalar_numeric}
\title{Check if object is scalar (length 1) numeric}
\usage{
is_scalar_numeric(x)
}
\arguments{
\item{x}{object to be checked}
}
\value{
scalar logical
}
\description{
Check if object is scalar (length 1) numeric
}
\seealso{
Other predicates: \code{\link{is_binary_valued}},
  \code{\link{is_numeric_like}},
  \code{\link{is_wholenumber}}
}
