% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{are_whole_number}
\alias{are_whole_number}
\alias{is_whole_number}
\alias{is_scalar_whole_number}
\title{Checks if an object is a whole number}
\usage{
are_whole_number(x, tol = .Machine$double.eps^0.5)

is_whole_number(x, tol = .Machine$double.eps^0.5, n = NULL, na_rm = FALSE)

is_scalar_whole_number(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{object to be checked}

\item{tol}{scalar numeric: tolerance within which a whole number will be
declared}

\item{n}{scalar whole number: expected length of vector}

\item{na_rm}{scalar logical: whether or not NAs should be considered in the
calculation. If na_rm = FALSE, then any NA in the vector being checked will
result in an NA output.}
}
\value{
scalar logical
}
\description{
Checks if an object is equal to its rounded form, minus some tolerance.
}
\section{Ignoring NAs}{
NA
}

\seealso{
Other predicates: \code{\link{are_non_negative}},
  \code{\link{are_numeric_like}},
  \code{\link{are_positive}}, \code{\link{are_zero}},
  \code{\link{is_binary_valued}},
  \code{\link{is_not_null}},
  \code{\link{is_scalar_bare_numeric}},
  \code{\link{is_zero}}
}
