% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{lst_combine}
\alias{lst_combine}
\title{Combines lists of vectors. Useful for generating lists of substitutions}
\usage{
lst_combine(x, method = c("stack", "merge", "cross"))
}
\arguments{
\item{x}{list of vectors}

\item{method}{scalar character. one of \code{c('stack', 'merge', 'cross')}.}
}
\value{
list
}
\description{
Combines lists of vectors. Useful for generating lists of substitutions
}
\section{Combination Method}{
 `cross` generates all pairwise combinations. `merge` combines element wise. `stack` generates one long list.
}

