% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzid.R
\name{mzids}
\alias{mzids}
\title{Creates a new mzids object}
\usage{
mzids(mz, rt, n_mz = 6, n_rt = 4, sep = "_", tag = "mzid",
  source = NA)
}
\arguments{
\item{mz}{vector numeric: The mz and rt. Scalar values are broadcast.}

\item{rt}{vector numeric: The mz and rt. Scalar values are broadcast.}

\item{n_mz}{scalar integerish: The number of digits after the decimal
point from the mz and the rt to be used in}

\item{n_rt}{scalar integerish: The number of digits after the decimal
point from the mz and the rt to be used in}

\item{sep}{scalar character: character separator in mzid}

\item{tag}{scalar character: tag at start of mzid}

\item{source}{vector character: data source of the mzids. Pass `NA` to keep
the field, but fill with `NA`, pass `NULL` to create an object without the
fields. Scalar values are broadcast}
}
\value{
mzid object
}
\description{
What is an mzids object and why would I want to use it?
}
