% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{neg_log}
\alias{neg_log}
\alias{neg_log_10}
\title{Helper functions to compute negative logs.}
\usage{
neg_log(x, base = exp(1))

neg_log_10(x)
}
\arguments{
\item{x}{vector numeric}

\item{base}{scalar numeric: the base with respect to which logarithms are
computed. Defaults to exp(1).}
}
\value{
vector numeric
}
\description{
Helper functions to compute negative logs.
}
