% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzid.R
\name{new_mzids}
\alias{new_mzids}
\title{Title}
\usage{
new_mzids(mz, rt, n_mz, n_rt, sep, tag, source = NA)
}
\arguments{
\item{mz, rt}{vector numeric: The mz and rt. Scalar values are broadcast.}

\item{n_mz, n_rt}{scalar integerish: The number of digits after the decimal
point from the mz and the rt to be used in}

\item{sep}{scalar character: character separator in mzid}

\item{tag}{scalar character: tag at start of mzid}

\item{source}{vector character: data source of the mzids. Pass `NA` to keep
the field, but fill with `NA`, pass `NULL` to create an object without the
fields. Scalar values are broadcast}
}
\value{
mzids object
}
\description{
Title
}
