% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\alias{\%<>\%}
\alias{.data}
\title{Pipe}
\arguments{
\item{lhs, rhs}{Pipe the output of \code{lhs} to the first input of \code{rhs}}

\item{lhs, rhs}{Pipe the output of \code{lhs} to the first input of \code{rhs}
and assign to the left hand side}
}
\description{
Use the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.

Use the double pipe function, \code{\%<>\%} to turn function composition into
a series of imperative statements.

Use the .data when writing functions that use piping with dplyr to avoid lazy
eval issues and get informative error messages
}
\examples{
1 \%>\% `+`(1)
a = 1
a \%<>\% `+`(1)
a
}
\keyword{internal}
