% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{probe}
\alias{probe}
\title{Internal version of `map_lgl()`` that works with logical vectors}
\usage{
probe(.x, .p, ...)
}
\arguments{
\item{.p}{predicate function}

\item{x}{lits or atomic vector}
}
\value{
vector logical of same length as x
}
\description{
Used for `_if` type functions
}
\keyword{internal}
