% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trunc2}
\alias{trunc2}
\title{Truncates a number to d decimal points}
\usage{
trunc2(x, d = 0)
}
\arguments{
\item{x}{vector numeric}

\item{d}{scalar integerish}
}
\value{
vector numeric
}
\description{
Truncates a number to d decimal points
}
