% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip.R
\name{zip}
\alias{zip}
\alias{zip_lgl}
\alias{zip_chr}
\alias{zip_int}
\alias{zip_dbl}
\alias{zip2}
\alias{zip2_lgl}
\alias{zip2_chr}
\alias{zip2_int}
\alias{zip2_dbl}
\title{Zip together lists of vectors ala python}
\usage{
zip(.l)

zip_lgl(.l)

zip_chr(.l)

zip_int(.l)

zip_dbl(.l)

zip2(.x, .y)

zip2_lgl(.x, .y)

zip2_chr(.x, .y)

zip2_int(.x, .y)

zip2_dbl(.x, .y)
}
\arguments{
\item{.l}{list of vectors}

\item{.x, .y}{vectors}
}
\value{
list
}
\description{
\code{zip} alwyas returns a lists of lists. \code{zip_} returns a lists of
vectors of the corresponding type.
}
\note{
zip unzips itself!
}
