% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzid.R
\name{as.data.frame.mzid}
\alias{as.data.frame.mzid}
\alias{as.data,frame}
\title{Title}
\usage{
\method{as.data.frame}{mzid}(x, ...)
}
\arguments{
\item{x}{any \R object.}

\item{...}{additional arguments to be passed to or from methods.}
}
\value{
data frame
}
\description{
Title
}
