% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dat_to_mat}
\alias{dat_to_mat}
\title{Converts a dataframe to a matrix where a column of the dataframe
becomes the rownames of the matrix}
\usage{
dat_to_mat(x, nm_col = 1)
}
\arguments{
\item{x}{data frame}

\item{nm_col}{int: which column should be removed and used as row names?}
}
\value{
matrix
}
\description{
Converts a dataframe to a matrix where a column of the dataframe
becomes the rownames of the matrix
}
\keyword{internal}
