% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{fo_to_char}
\alias{fo_to_char}
\alias{as.character.formula}
\alias{as.character}
\title{fo_to_char}
\usage{
fo_to_char(x, ...)

\method{as.character}{formula}(x, ...)
}
\arguments{
\item{x}{formula}

\item{...}{arguments passed to and from other methods.}
}
\value{
A character vector
}
\description{
Converts a formula to character representaion.
}
\details{
`fo_to_char`` is a function provided for use to interface with
`as.character.formula`. If 'biodatacoreUtils' is loaded, then `as.character`
should automatically find the `.formula` method. `deparse_to_str` is the
underlying function that does the conversion.
}
