% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{inv_which}
\alias{inv_which}
\title{Given an object and a character or integerish vector, return a logical vctor
of positions.}
\usage{
inv_which(x, sel)
}
\arguments{
\item{x}{lits or atomic vector}

\item{sel}{A character vector of names or a numeric vector of positions}
}
\value{
vector logical of same length as x
}
\description{
Used for `_at` type functions.
}
\keyword{internal}
