% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_binary_valued}
\alias{is_binary_valued}
\title{Check if a vector has only two unique entries}
\usage{
is_binary_valued(x, n = NULL, na_rm = FALSE)
}
\arguments{
\item{x}{object to be checked}

\item{n}{scalar whole number: expected length of vector}

\item{na_rm}{scalar logical: whether or not NAs should be considered in the
calculation. If na_rm = FALSE, then any NA in the vector being checked will
result in an NA output.}
}
\value{
scalar logical
}
\description{
Check if a vector has only two unique entries
}
\section{Ignoring NAs}{
NA
}

\seealso{
Other predicates: \code{\link{are_non_negative}},
  \code{\link{are_numeric_like}},
  \code{\link{are_positive}},
  \code{\link{are_whole_number}}, \code{\link{are_zero}},
  \code{\link{is_not_null}},
  \code{\link{is_scalar_bare_numeric}},
  \code{\link{is_zero}}
}
