% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{are_numeric_like}
\alias{are_numeric_like}
\alias{is_numeric_like}
\alias{is_scalar_numeric_like}
\title{Checks if a an object can be converted to a number without error}
\usage{
are_numeric_like(x)

is_numeric_like(x, n = NULL, na_rm = FALSE)

is_scalar_numeric_like(x)
}
\arguments{
\item{x}{object to be checked}

\item{n}{scalar whole number: expected length of vector}

\item{na_rm}{scalar logical: whether or not NAs should be considered in the
calculation. If na_rm = FALSE, then any NA in the vector being checked will
result in an NA output.}
}
\value{
scalar logical
}
\description{
Checks if a an object can be converted to a number without error
}
\section{Ignoring NAs}{
NA
}

\seealso{
Other predicates: \code{\link{are_non_negative}},
  \code{\link{are_positive}},
  \code{\link{are_whole_number}}, \code{\link{are_zero}},
  \code{\link{has_zero_range}},
  \code{\link{is_binary_valued}},
  \code{\link{is_not_null}},
  \code{\link{is_scalar_bare_numeric}},
  \code{\link{is_zero}}

Other predicates: \code{\link{are_non_negative}},
  \code{\link{are_positive}},
  \code{\link{are_whole_number}}, \code{\link{are_zero}},
  \code{\link{has_zero_range}},
  \code{\link{is_binary_valued}},
  \code{\link{is_not_null}},
  \code{\link{is_scalar_bare_numeric}},
  \code{\link{is_zero}}
}
