% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_wholenumber}
\alias{is_wholenumber}
\alias{is_scalar_wholenumber}
\title{Checks if an object is/contains only whole numbers}
\usage{
is_wholenumber(x, tol = .Machine$double.eps^0.5)

is_scalar_wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{object to be checked.}

\item{tol}{tolerance within which a whole number will be declared}
}
\value{
scalar logical
}
\description{
Checks if an object is/contains only whole numbers
}
\seealso{
Other predicates: \code{\link{is_binary_valued}},
  \code{\link{is_numeric_like}},
  \code{\link{is_scalar_numeric}}
}
