% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicates.R
\name{is_mzid}
\alias{is_mzid}
\alias{is_mzids}
\title{Title}
\usage{
is_mzid(x)

is_mzids(x)
}
\arguments{
\item{x}{mzid object}

\item{x}{mzid object}
}
\description{
Title

Title
}
\section{Functions}{
\itemize{
\item \code{is_mzid}: 

\item \code{is_mzids}: 
}}

